requirements.txt
altair==5.5.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
firebase_admin==7.1.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.7
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
MarkupSafe==3.0.2
msgpack==1.1.1
narwhals==2.0.1
numpy==2.3.2
oauth2client==4.1.3
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
